requirements.txt
Django==2.2.3
django-appconf==1.0.3
django-crispy-forms==1.7.2
gunicorn==19.9.0
Pillow==6.1.0
psycopg2-binary==2.8.3
pytz==2019.1
rope==0.14.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
